scripts/requirements-vector.txt
lancedb>=0.4
sentence-transformers>=2.2
pyarrow>=14.0
pyyaml>=6.0
